[REFACTO] Cypress folder refacto
