+++ cypress/e2e/test_calculus_graph.cy.js
