+++ cypress/e2e/test_result_panel.cy.js
@@ -1,6 +1,31 @@
 import "cypress-real-events";
-describe("Test - Import JSON", () => {
-    it("Pick one JSON file and check trasnfert status", () => {
+describe("Test - Result panel", () => {
+    it("Check if the model cannot be calculated if the modal exception is displayed and the result panel not showed", () => {
+        let ujName = "Test E2E UJ";
+        let ujsOne = "Test E2E UJ 1";
+        let ujsTwo = "Test E2E UJ 2";
+        let server = "Test E2E Server";
+        let service = "Test E2E Service";
+        let jobOne = "Test E2E Job 1";
+        let jobTwo = "Test E2E Job 2";
+        let upName = "Test E2E Usage Pattern";
+
+        cy.visit("/model_builder/");
+        cy.get('button[hx-get="/model_builder/open-import-json-panel/"]').click();
+        let fileTest = 'cypress/fixtures/efootprint-model-no-job.json'
+        cy.get('input[type="file"]').selectFile(fileTest);
+        cy.get('button[type="submit"]').click();
+
+        cy.get('#btn-open-panel-result')
+        .realTouch('start', { x: 100, y: 300 })
+        .realTouch('move', { x: 100, y: 200 })
+        .realTouch('end', { x: 100, y: 200 });
+
+        cy.get('#panel-result-btn').should('not.have.css', 'height', '93vh');
+        cy.get('button').contains('Go back').should('be.exist');
+    });
+
+    it("Check if the model is can be calculated the panel is displayed and can be swipe down", () => {
         let upName = "Test E2E Usage Pattern";
 
         cy.visit("/");
